Fixed for TPU.

+++ src/agent/encoder.py
@@ -17,7 +17,7 @@ class Encoder(nn.Module):
     def forward(self, src, real_positions, eos_positions):
         src = src.transpose(0, 1)
         # eos_positions = eos_positions.transpose(0, 1).unsqueeze(-1)
-        att_add_mask = torch.log(real_positions)
+        att_add_mask = torch.log(real_positions.float())
 
         # eos_value = eos_positions * src
         src = src + self.pos_encoder(src)  # * math.sqrt(Config.vector_sizes[level])
